# frontend/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CodigoActivo API",
    "description": "Backend API for CodigoActivo (cookie session auth + CSRF).",
    "version": "v1"
  },
  "paths": {
    "/api/activities/{eventId}/household-assignments": {
      "get": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdMemberAssignmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/{activityId}/{userId}/verifyTimeOverlaps": {
      "get": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeOverlapResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/{eventId}": {
      "post": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/{activityId}": {
      "put": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/activities/{activityId}/{userId}/assign": {
      "patch": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/{activityId}/assign-household": {
      "post": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignHouseholdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/{activityId}/{userId}/unassign": {
      "patch": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/activities/{activityId}/{userId}/change-status": {
      "patch": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAssignmentStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/{activityId}/{userId}/change-role": {
      "patch": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAssignmentRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/roleType": {
      "post": {
        "tags": [
          "Activities"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityRoleTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityRoleTypeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/activities/roleType/{activityRoleTypeId}": {
      "put": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityRoleTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityRoleTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityRoleTypeResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Activities"
        ],
        "parameters": [
          {
            "name": "activityRoleTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/announcements": {
      "post": {
        "tags": [
          "Announcements"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/announcements/{announcementId}": {
      "put": {
        "tags": [
          "Announcements"
        ],
        "parameters": [
          {
            "name": "announcementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnouncementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Announcements"
        ],
        "parameters": [
          {
            "name": "announcementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/announcements/{announcementId}/feature": {
      "patch": {
        "tags": [
          "Announcements"
        ],
        "parameters": [
          {
            "name": "announcementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityAllowedRoleRequest": {
        "type": "object",
        "properties": {
          "activityRoleTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "desiredSignups": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ActivityAllowedRoleResponse": {
        "type": "object",
        "properties": {
          "roleTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "roleTypeName": {
            "type": "string",
            "nullable": true
          },
          "desiredSignups": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ActivityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "activityStartsAt": {
            "type": "string",
            "format": "date-time"
          },
          "activityEndsAt": {
            "type": "string",
            "format": "date-time"
          },
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "updatedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "allowedRoleTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityAllowedRoleResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ActivityRoleTypeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AnnouncementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "updatedBy": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          },
          "featured": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AssignHouseholdRequest": {
        "type": "object",
        "properties": {
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HouseholdAssignmentRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AssignRequest": {
        "type": "object",
        "properties": {
          "activityRoleTypeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "AssignmentResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "roleTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "roleTypeName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/AssignmentStatusResponse"
          }
        },
        "additionalProperties": false
      },
      "AssignmentStatusResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChangeAssignmentRoleRequest": {
        "type": "object",
        "properties": {
          "activityRoleTypeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "ChangeAssignmentStatusRequest": {
        "type": "object",
        "properties": {
          "assignmentStatusId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CreateActivityRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "activityStartsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "activityEndsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          },
          "allowedRoleTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityAllowedRoleRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateActivityRoleTypeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateAnnouncementRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "HouseholdAssignmentRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "activityRoleTypeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "HouseholdMemberAssignmentResponse": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "roleTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "statusId": {
            "type": "string",
            "format": "uuid"
          },
          "statusName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OverlappingActivityResponse": {
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "TimeOverlapResponse": {
        "type": "object",
        "properties": {
          "hasOverlaps": {
            "type": "boolean"
          },
          "overlaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OverlappingActivityResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateActivityRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "activityStartsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "activityEndsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          },
          "allowedRoleTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityAllowedRoleRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateActivityRoleTypeRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateAnnouncementRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "thumbnailId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}